Main cost function for four parameter optimization for the utterance choice experiments. Optimizing softness, non-obedience, alpha and klValueFactor.

Full-RSA

The actual RSA model Kullback-Leibler divergence determination for utterance choice experiments.

Softness, non-obedience, alpha and klValueFactor are optimized.

RSAModelUttKLDiv_4params(data, par1, par2, par3, par4)

Arguments

data
A matrix with data rows.

column structure: [1:OC1,OC2,OC3,4:numUttOptions,7-X:TurkerSliderValues]

1:OC1 Object 1. A value between 1 and 27.

2:OC2 Object 2. A value between 1 and 27.

3:OC3 Object 3. A value between 1 and 27.

4:numUttOptions The number of valid utterances in the scene.

7-X:TurkerSliderValues These columns contain the participants' slider values.

par1

softness parameter

A parameter value between [0,infinity) (The larger the value the higher the tendency towards uniform liking).Value reflects how categorical the listener's preferences are:0: The listener always picks her preferred object.If the listener prefers red objects, she will always pick the red object in the scene.infinity: It is as likely for the listener to pick green, blue or red objects.

par2

non-obedience parameter

Determines the extent to which the instruction of the speaker is obeyed by the listener.(0 = full obedience, infinity = full instruction ignorance).Example:0: Listener always picks red objects following the utterance "red".infinity: Listener as likely to pick green, blue or red objects even if the utterance is "red".

par3

alpha parameter

A parameter value between 0 and 1.This parameter is an exponential scaling of the speaker choosing the utterance that maximizes the chance of the listener getting the target object right.

par4
klValueFactor can be negative, 0 or positive:

zero

Don't care about learning about the feature preferences of the listener

positive

Care about learning about the feature preferences of the listener

negative

Trying to pick non-ambiguous utterances

Value

Minimized Kullback-Leibler divergence and the optimal parameters.

Details

This function uses getBestInfGainUttPreferences.

And is used by

RSAModelUttKLDivParamD4,

RSAModelUttKLDivParamD4A.2,

RSAModelUttKLDivParamD4B.2,

RSAModelUttKLDivParamD4AB.2,

RSAModelUttKLDivParamAD4,

RSAModelUttKLDivParamBD4,

RSAModelUttKLDivParamAD4B.2,

RSAModelUttKLDivParamBD4A.2,

RSAModelUttKLDivParamCD4,

RSAModelUttKLDivParamBCD4,

RSAModelUttKLDivParamACD4,

RSAModelUttKLDivParamABD4,

RSAModelUttKLDivParamABCD4.