Uniform Model Kullback-Leibler divergence determination

Full-RSA

Divergence of the observed distribution from the uniform distribution over preferences.

UniformModelKLDiv(data)

Arguments

data
A matrix with data rows.

column structure:

[1:OC1,OC2,OC3,4:UUFeat, 5:Q1Feat,6:Q2Feat]

[7:Q1AnswerV1,V2,V3, 10:Q2AnswerV1,V2,V3]

1:OC1 Object 1. A value between 1 and 27.

2:OC2 Object 2. A value between 1 and 27.

3:OC3 Object 3. A value between 1 and 27.

4:UUFeat Uttered feature. A number between 1 and 3. (1: shape, 2: pattern, 3: color)

5:Q1Feat Questioned feature 1. A number between 1 and 3. (1: shape, 2: pattern, 3: color).

Example: If you utter "blue" (feature: color), then you can learn something about shape and texture preferences.

6:Q2Feat Questioned feature 2. A number between 1 and 3. (1: shape, 2: pattern, 3: color).

Example: If you utter "blue" (feature: color), then you can learn something about shape and texture preferences.

7:Q1AnswerV1, V2, V3 The columns 7-9 contain the participants' slider values for the first questioned feature.

10:Q2AnswerV1, V2, C3 The columns 10-12 contain the participants' slider values for the second questioned feature.

Value

Minimized Kullback-Leibler divergence and the optimal parameters.